video_uploader/requirements.txt
fastapi==0.88.0
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.1
uvicorn==0.20.0
python-multipart==0.0.5
pika==1.3.1